+++ app/javascript/readingList/components/ItemListItem.jsx
